Writing json files to the correct path for Earth
The Earth visualization needs to read the weather data in Year, Month,
and Day directories, so we look at the date argument and then create the
resulting directories.

I also added a function to define a unique location for the resulting
data file in case we need to use this for a one off process.